Changes here and here: 1. added connecttime mesaurement test test/connecttime/ 2. started including multipaxos/ in GNS code 3. Hostname to IP address is lookup upon first access. and not immediately after starting a node. ... some more .. can't remeber
git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@739 0b15c3b6-4604-4ca4-ad59-2e1af984e93c

+++ src/edu/umass/cs/gns/nsdesign/gnsReconfigurable/StaticReplicationCoordinator.java
@@ -62,7 +62,7 @@ public class StaticReplicationCoordinator extends ActiveReplicaCoordinator{
     HashMap<String, Set<Integer>> groupIDsMembers = ConsistentHashing.getReplicaControllerGroupIDsForNode(nodeID);
     for (String groupID : groupIDsMembers.keySet()) {
       GNS.getLogger().info("Creating paxos instances: " + groupID + "\t" + groupIDsMembers.get(groupID));
-      paxosManager.createPaxosInstance(groupID, 1, groupIDsMembers.get(groupID), paxosInterface);
+      paxosManager.createPaxosInstance(groupID, (short) 1, groupIDsMembers.get(groupID), paxosInterface);
     }
   }
